+++ plone/releaser/release.py
@@ -50,22 +50,8 @@ def check_pypi_access(data):
 
 
 def show_changelog_entries(data):
-    # Find changelog
-    # TODO: Figure out how to catch malformed rst
-    if data['history_file'] is not None:
-        changelog = Changelog(file_location=data['history_file'])
-    # Get top release's entry
-    entries = changelog.latest()
-    if entries is None:
-        if not ask("Unable to parse changelog. Continue?", default=True):
-            sys.exit()
-        return
     print "Changelog entries for version {0}.".format(data['new_version'])
-    for entry in entries:
-        if isinstance(entry, list):
-            print '\n'.join(entry)
-        else:
-            print entry
+    print data.get('history_last_release')
     if not ask("Continue?", default=True):
         sys.exit()